Dividindo em bibliotecas
Reparti todas as funções em 3 bibliotecas:

menu.h/menu.c: Funções do menu.
game.h/game.c: Funções para funcionamento geral do jogo.
records.h/records.c: Funções para salvar os resultados das partidas.
